backend/docx_to_pdf_app/requirements.txt
Flask==3.0.0
python-docx==1.1.0
reportlab==4.0.4
Pillow==10.1.0
werkzeug==3.0.1